requirements.txt
absl-py==1.3.0
# aclpubcheck==0.1
alive-progress==1.6.2
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
bibtexparser==1.4.0
black==22.6.0
bleach==4.1.0
blis==0.7.5
Bottleneck==1.3.5
brotlipy==0.7.0
catalogue==2.0.6
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.5
commonmark==0.9.1
cryptography==38.0.1
cycler==0.11.0
cymem==2.0.6
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
# en-core-web-lg==3.0.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.4.2
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.1.0
idna==3.4
importlib-metadata==4.11.3
importlib-resources==5.2.0
ipykernel==6.15.2
ipympl==0.9.2
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.1
json5==0.9.6
jsonschema==4.16.0
jupyter_client==7.3.5
jupyter_core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.2.8
jupyterlab-pygments==0.1.2
jupyterlab_server==2.15.2
jupyterlab-widgets==3.0.3
keras==2.7.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
latexcodec==2.0.1
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.2.2
matplotlib-inline==0.1.6
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
murmurhash==1.0.6
mypy-extensions==0.4.3
nbclassic==0.3.5
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.5.0
nest-asyncio==1.5.5
nltk==3.6.7
notebook==6.4.6
numexpr==2.8.3
numpy==1.23.3
olefile==0.46
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathy==0.6.1
patsy==0.5.2
pdfminer.six==20221105
pdfplumber==0.7.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==22.2.2
pipreqs==0.4.11
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
plotly==5.11.0
preshed==3.0.6
prettytable==3.6.0
prometheus-client==0.14.1
prompt-toolkit==3.0.20
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1-modules==0.2.8
pybtex==0.24.0
pycparser==2.21
pydantic==1.7.4
Pygments==2.11.2
pylatexenc==2.10
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.0
PySocks==1.7.1
python-crfsuite==0.9.7
python-dateutil==2.8.2
pytorch-ignite==0.4.10
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
rebiber==1.1.3
regex==2022.7.9
requests==2.28.1
rich==12.6.0
sacremoses==0.0.47
scikit-learn==1.1.2
scikit-learn-extra==0.2.0
scipy==1.9.1
seaborn==0.12.0
Send2Trash==1.8.0
sentence-transformers==1.1.0
sentencepiece==0.1.96
setuptools==65.5.0
six==1.16.0
sklearn-crfsuite==0.3.6
smart-open==5.2.1
sniffio==1.2.0
soupsieve==2.3.2.post1
spacy==3.0.5
spacy-legacy==3.0.8
srsly==2.4.2
stack-data==0.2.0
statsmodels==0.13.2
tabulate==0.9.0
tenacity==8.0.1
tensorboard==1.15.0
tensorboard-data-server==0.6.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.6.0
thinc==8.0.13
threadpoolctl==3.0.0
tokenizers==0.10.3
tomli==2.0.1
torch==1.10.2
tornado==6.2
tqdm==4.64.0
traitlets==5.1.1
transformers==4.5.1
tsv==1.2
typer==0.3.2
typing_extensions==4.3.0
Unidecode==1.3.6
urllib3==1.26.12
Wand==0.6.11
wasabi==0.9.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.0.3
wheel==0.37.1
widgetsnbextension==4.0.3
wrapt==1.12.1
yarg==0.1.9
zipp==3.8.0